+++ package.json
@@ -73,9 +73,9 @@
     "@storybook/react": "^6.3.9",
     "@storybook/source-loader": "^6.3.9",
     "@testing-library/dom": "^8.7.2",
-    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.2.1",
+    "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.0.2",
     "@types/jest-axe": "^3.5.3",
     "@types/node": "^16.10.2",
