+++ src/js/components/Carousel/__tests__/Carousel-test.tsx
@@ -1,139 +1,167 @@
 import React from 'react';
-import { render, fireEvent, act } from '@testing-library/react';
+import { render, act, waitFor, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import 'jest-styled-components';
+import '@testing-library/jest-dom';
 
-import { Grommet } from '../../Grommet';
 import { Carousel } from '..';
-import { Image } from '../../Image';
+
+const getSlideOne = () => screen.getByText('Slide One');
+const getSlideTwo = () => screen.getByText('Slide Two');
+const getSlideThree = () => screen.getByText('Slide Three');
 
 describe('Carousel', () => {
   test('basic', () => {
     const { container } = render(
-      <Grommet>
-        <Carousel>
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
+      <Carousel>
+        <div>Slide One</div>
+        <div>Slide Two</div>
+      </Carousel>,
     );
 
+    expect(getSlideOne()).toHaveStyle({ visibility: 'visible' });
+    expect(getSlideTwo()).toHaveStyle({ visibility: 'hidden' });
     expect(container.firstChild).toMatchSnapshot();
   });
 
   test('basic with `initialChild: 1`', () => {
     const { container } = render(
-      <Grommet>
-        <Carousel initialChild={1}>
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
+      <Carousel initialChild={1}>
+        <div>Slide One</div>
+        <div>Slide Two</div>
+      </Carousel>,
     );
 
+    expect(getSlideOne()).toHaveStyle({ visibility: 'hidden' });
+    expect(getSlideTwo()).toHaveStyle({ visibility: 'visible' });
     expect(container.firstChild).toMatchSnapshot();
   });
 
-  test('navigate', () => {
-    const { getByTestId, container } = render(
-      <Grommet>
-        <Carousel data-testid="test-carousel">
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.keyDown(getByTestId('test-carousel'), {
-      key: 'Right',
-      keyCode: 39,
-      which: 39,
-    });
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.keyDown(getByTestId('test-carousel'), {
-      key: 'Left',
-      keyCode: 37,
-      which: 37,
-    });
-    expect(container.firstChild).toMatchSnapshot();
-  });
-
-  test('should trigger events of focus, blur and click', () => {
-    const onFocus = jest.fn();
-    const onBlur = jest.fn();
-    const { container } = render(
-      <Grommet>
-        <Carousel
-          id="test-carousel"
-          onFocus={onFocus}
-          onBlur={onBlur}
-          controls="selectors"
-        >
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
-    );
-
-    const button = document
-      .getElementById('test-carousel')!
-      .querySelector('button');
-
-    expect(button).not.toBeNull();
-
-    fireEvent.focus(button!);
-    expect(onFocus).toHaveBeenCalledTimes(1);
-
-    fireEvent.click(button!);
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.blur(button!);
-    expect(onBlur).toHaveBeenCalledTimes(1);
-  });
-
-  test('play', () => {
-    jest.useFakeTimers('modern');
-    const { container } = render(
-      <Grommet>
-        <Carousel play={1000}>
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    // Advance timers so the carousel advances
-    act(() => {
-      jest.advanceTimersByTime(1300);
-    });
-    expect(container.firstChild).toMatchSnapshot();
-  });
-
-  test('should work as controlled carousel', () => {
-    const setActiveSlide = jest.fn();
-
+  test('arrow navigation: next', async () => {
     render(
-      <Grommet>
-        <Carousel
-          id="test-carousel"
-          controls="selectors"
-          activeChild={1}
-          onChild={setActiveSlide}
-        >
-          <Image src="//v2.grommet.io/assets/IMG_4245.jpg" />
-          <Image src="//v2.grommet.io/assets/IMG_4210.jpg" />
-        </Carousel>
-      </Grommet>,
+      <Carousel fill controls="arrows">
+        <div>Slide One</div>
+        <div>Slide Two</div>
+        <div>Slide Three</div>
+      </Carousel>,
     );
 
-    const button = document
-      .getElementById('test-carousel')!
-      .querySelector('button');
+    /**
+     * - Currently on "Slide One"
+     * - Simulating click on the next arrow button
+     * - Expecting "Slide Two" to be visible
+     */
+    const nextButton = screen.getByLabelText('Next');
+    expect(getSlideTwo()).toHaveStyle({ visibility: 'hidden' });
+    userEvent.click(nextButton);
+    await waitFor(() => {
+      expect(getSlideTwo()).toHaveStyle({ visibility: 'visible' });
+    });
+  });
 
-    expect(button).not.toBeNull();
+  test('arrow navigation: previous', async () => {
+    render(
+      <Carousel fill controls="arrows">
+        <div>Slide One</div>
+        <div>Slide Two</div>
+        <div>Slide Three</div>
+      </Carousel>,
+    );
 
-    fireEvent.click(button!);
-    expect(setActiveSlide).toBeCalledWith(0);
+    /**
+     * - Currently on "Slide One"
+     * - Simulating click on the previous arrow button
+     * - Expecting "Slide Three" to be visible
+     */
+    const previousButton = screen.getByLabelText('Previous');
+    expect(getSlideThree()).toHaveStyle({ visibility: 'hidden' });
+    userEvent.click(previousButton);
+    await waitFor(() => {
+      expect(getSlideThree()).toHaveStyle({ visibility: 'visible' });
+    });
+  });
+
+  test('selector navigation: forward', async () => {
+    render(
+      <Carousel fill controls="selectors">
+        <div>Slide One</div>
+        <div>Slide Two</div>
+        <div>Slide Three</div>
+      </Carousel>,
+    );
+
+    /**
+     * - Currently on "Slide One"
+     * - Simulating click on the 3rd selector button
+     * - Expecting "Slide Three" to be visible
+     */
+    const thirdSelector = screen.getAllByRole('button')[2];
+    expect(getSlideThree()).toHaveStyle({ visibility: 'hidden' });
+    userEvent.click(thirdSelector);
+    await waitFor(() => {
+      expect(getSlideThree()).toHaveStyle({ visibility: 'visible' });
+    });
+  });
+
+  test('selector navigation: backward', async () => {
+    render(
+      <Carousel initialChild={2} fill controls="selectors">
+        <div>Slide One</div>
+        <div>Slide Two</div>
+        <div>Slide Three</div>
+      </Carousel>,
+    );
+
+    /**
+     * - Currently on "Slide Three"
+     * - Simulating click on the 1st selector button
+     * - Expecting "Slide One" to be visible
+     */
+    const firstSelector = screen.getAllByRole('button')[0];
+    expect(getSlideOne()).toHaveStyle({ visibility: 'hidden' });
+    userEvent.click(firstSelector);
+    await waitFor(() => {
+      expect(getSlideOne()).toHaveStyle({ visibility: 'visible' });
+    });
+  });
+
+  test('play', async () => {
+    jest.useFakeTimers('modern');
+    render(
+      <Carousel play={500} controls={false}>
+        <div>Slide One</div>
+        <div>Slide Two</div>
+      </Carousel>,
+    );
+
+    expect(screen.getByText('Slide One')).toHaveStyle({
+      visibility: 'visible',
+    });
+    expect(screen.getByText('Slide Two')).toHaveStyle({
+      visibility: 'hidden',
+    });
+
+    await act(async () => {
+      jest.advanceTimersByTime(800);
+    });
+    await waitFor(() => {
+      expect(getSlideOne()).toHaveStyle({ visibility: 'hidden' });
+    });
+    await waitFor(() => {
+      expect(getSlideTwo()).toHaveStyle({ visibility: 'visible' });
+    });
+  });
+
+  test('controlled component', async () => {
+    const setActiveSlide = jest.fn();
+    const { container } = render(
+      <Carousel controls="arrows" activeChild={1} onChild={setActiveSlide}>
+        <div>Slide One</div>
+        <div>Slide Two</div>
+      </Carousel>,
+    );
+
+    expect(container.firstChild).toMatchSnapshot();
+    expect(setActiveSlide).toHaveBeenCalledWith(1);
   });
 });
