+++ src/js/components/Carousel/stories/Autoplay.js
@@ -1,11 +1,15 @@
 import React from 'react';
 import { Attraction, Car, TreeOption } from 'grommet-icons';
-import { Box } from 'grommet';
+
+import { Text, Box } from 'grommet';
 import { Carousel } from '../Carousel';
 
-export const Simple = () => (
-  <Box align="center" pad="large">
-    <Carousel>
+export const Autoplay = () => (
+  <Box gap="medium" align="center" pad="large">
+    <Text size="small">
+      The Carousel slides will transition every 3 seconds
+    </Text>
+    <Carousel controls={false} play={3000}>
       <Box pad="xlarge" background="accent-1">
         <Attraction size="xlarge" />
       </Box>
@@ -19,10 +23,6 @@ export const Simple = () => (
   </Box>
 );
 
-Simple.parameters = {
-  chromatic: { disable: true },
-};
-
 export default {
-  title: 'Media/Carousel/Simple',
+  title: 'Media/Carousel/Autoplay',
 };
